% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bayesQR.R
\name{bayesQR}
\alias{bayesQR}
\title{Bayesian quantile regression model}
\usage{
bayesQR(formula, tau = 0.5, data, itNum, thin = 1, betaValue = NULL,
  sigmaValue = 1, priorVar = 100, refresh = 100)
}
\arguments{
\item{formula}{a formula object, with the response on the left of a ~
operator, and the terms, separated by + operators, on the right.}

\item{tau}{Quantile of interest.}

\item{data}{A data.frame from which to find the variables defined in the
formula}

\item{itNum}{Number of iterations.}

\item{thin}{Thinning parameter. Default value is 1.}

\item{betaValue}{Initial values for the parameter beta for the continuous
part.}

\item{sigmaValue}{Initial value for the scale parameter.}

\item{priorVar}{Value that multiplies a identity matrix in the elicition
process of the prior variance of the regression parameters.}

\item{refresh}{Interval between printing a message during the iteration
process. Default is set to 100.}

\item{gammaValue}{Initial value for the parameter gamma of the discrete
part.}
}
\value{
A list with the chains of all parameters of interest.
}
\description{
This function estimates a Bayesian quantile regression model
The response variable is assumed to follow a asymmetric Laplace
distribution.
}
\examples{
set.seed(1)
}
\references{
Kozumi and Kobayashi (2015) - Gibbs sampling methods for
Bayesian quantile regression.
}

