% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BrazilElection2014}
\alias{BrazilElection2014}
\title{Results of the Brazilian presidential election in 2014, with percentage of
 votes for Dilma Rousseff by city.}
\format{A data frame with 5501 rows and 13 variables:
\describe{
  \item{city}{Name of the city.}
  \item{UF}{Code for the state.}
  \item{UF_name}{Code of two letters for the state.}
  \item{Region}{Region of the city: 1 = North, 2 = Northeast,
  3 = Southeast, 4 = South, 5 = Midwest.}
  \item{income_delta}{Relative difference between income per capita in 2010
   and 2000.}
  \item{Gini_delta}{Relative difference between the Gini index in 2010 and
   2000.}
  \item{Years_educ}{Years of education.}
  \item{HDI}{Human development index.}
  \item{avepay_BF}{Average payment received for the participant families in
   the federal program for indirect cash transfer, called Bolsa-Familia.}
  \item{population}{Population number in 2010.}
  \item{latitude}{Latitude.}
  \item{longitude}{Longitude.}
}}
\source{
\url{}
}
\usage{
BrazilElection2014
}
\description{
A dataset containing data about the presidential election of 2014 in the
 total 5501 cities, where a few cities were not included in this dataset,
 for which it was not possible to match the information for votes and the
 other predictor variables.
}
\keyword{datasets}

