% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BrazilElectricity}
\alias{BrazilElectricity}
\title{Proportion of households with access to electricity in Brazil}
\format{A data frame with 500 rows and 6 variables:
\describe{
  \item{prop_elec}{propotion of households with access to electricity.}
  \item{region}{2 for Northeast, 3 for Southeast.}
  \item{population}{Population, in years.}
  \item{income_percap}{Income per capita, in Brazilian reais.}
  \item{hdi}{Human development index.}
  \item{pop_density}{Population density.}
}}
\source{
\url{http://www.ibge.gov.br/english/estatistica/populacao/
default_censo_2000.shtm}
}
\usage{
BrazilElectricity
}
\description{
A dataset containing a sample of cities in Brazil, with the proportion of
households with access to electricity with data from the Brazilian census
of 2000, and some sociodemographic variables.
}
\references{
Santos and Bolfarine (2015). Bayesian analysis for zero-or-one
 inflated proportion data using quantile regression. \emph{Journal of
 Statistical Computation and Simulation}. \url{http://www.tandfonline.com
 /doi/abs/10.1080/00949655.2014.986733}
}
\keyword{datasets}
