% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BrazilGini}
\alias{BrazilGini}
\title{Gini index for Brazilian states in three censuses.}
\format{A data frame with 81 rows and 4 variables:
\describe{
  \item{UFN}{Name of the state}
  \item{GINI}{Gini index}
  \item{INCPC}{Average income per capita}
  \item{EDUC}{Average years of education}
  \item{YEAR}{Year of the census}
}}
\source{
\url{http://www.ibge.gov.br/english/estatistica/populacao/
default_censo_2000.shtm}
}
\usage{
BrazilGini
}
\description{
A dataset containing data about the Gini index of 27 Brazilian states for
 three different censuses, 1991, 2000 and 2010, where each row represents
 the data about a specific state in a specific census year.
}
\references{
Santos and Bolfarine (2016). On Bayesian quantile regression
 and outliers. \url{http://arxiv.org/abs/1601.07344}
}
\keyword{datasets}

