% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkpoints_qreg_2D.R
\name{checkpoints_qreg_2D}
\alias{checkpoints_qreg_2D}
\title{Multiple-output Bayesian quantile regression model}
\usage{
checkpoints_qreg_2D(
  model,
  datafile,
  response,
  points_y,
  x_values = 1,
  path_folder = NULL,
  splines_part = FALSE,
  w_values = NULL,
  model_name = "bayesx.estim",
  name_var,
  lambda_a,
  ...
)
}
\arguments{
\item{model}{This is an object of the class \code{multBQR}, produced by a
call to the \code{multBayesQR} function.}

\item{datafile}{A data.frame from which to find the variables defined in the
formula.}

\item{response}{Names of response variables}

\item{points_y}{the exact points in Y, in which one wants to find its
respective quantile region.}

\item{x_values}{Fixed value of the predictor variables.}

\item{path_folder}{The path where all results are stored.}

\item{splines_part}{Logical value to indicate whether there are splines
terms in the equation to draw the quantile contours.}

\item{w_values}{Value to be considered in the nonlinear part of the model.}

\item{model_name}{When results will be collected in a folder, this should be
the name of the name considered by BayesX to save all tables. Default is
'bayesx.estim'.}

\item{name_var}{When there is a nonlinear variable from which one wants to
consider different values for plotting, this should have the name of the
variable.}

\item{lambda_a}{Parameter to be make the adjustment of the intercept to
correct the miscoverage of the quantile regions.}

\item{...}{Other parameters for \code{summary.multBQR}.}
}
\value{
A ggplot with the quantile regions based on Bayesian quantile
 regression model estimates.
}
\description{
This function checks whether points belong to quantile regions or not, based
 on estimated models for models with 2 dimensions.
}
