% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawQuantileRegion.R
\name{drawQuantileRegion}
\alias{drawQuantileRegion}
\title{Multiple-output Bayesian quantile regression model}
\usage{
drawQuantileRegion(model, ngridpoints = 100, xValue, paintedArea = TRUE,
  ...)
}
\arguments{
\item{model}{This is an object of the class \code{multBQR}, produced by a call to the
\code{multBayesQR} function.}

\item{ngridpoints}{Number of grid points considered to build this quantile region,
where a thorough search will look for the specified region, given the estimates
for several directions. Default is 100, that will produce a grid with 10.000 points
in the observed range of the data.}

\item{xValue}{Fixed value of the predictor variables. It is not necessary for models
with only the intercept.}

\item{paintedArea}{If TRUE, it will plot the data points and the quantile region layer
over the points, for each tau, in different plots. If FALSE, it will produce one
plot showing showing all quantile regions for the different quantiles.}
}
\value{
A ggplot with the quantile regions based on Bayesian quantile regression
model estimates.
}
\description{
This function draws plots of the quantile region based on multiple-output quantile
regression models.
}
