% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawQuantileRegion_3D.R
\name{drawQuantileRegion_3D}
\alias{drawQuantileRegion_3D}
\title{Multiple-output Bayesian quantile regression model}
\usage{
drawQuantileRegion_3D(model, datafile, response, ngridpoints = 100,
  xValue = 1, path_folder = NULL, splines_part = FALSE,
  wValue = NULL, print_plot = TRUE, model_name = "bayesx.estim",
  name_var, ...)
}
\arguments{
\item{model}{This is an object of the class \code{multBQR}, produced by a
call to the \code{multBayesQR} function.}

\item{datafile}{A data.frame from which to find the variables defined in the
formula.}

\item{response}{Names of response variables}

\item{ngridpoints}{Number of grid points considered to build this quantile
region, where a thorough search will look for the specified region, given
the estimates for several directions. Default is 100, that will produce a
grid with 10.000 points in the observed range of the data.}

\item{xValue}{Fixed value of the predictor variables. Default value is 1
when there is only the intercept. If there is the interest in comparing the
quantile regions for different values of predictors, it must used with a
list of values for the predictors.}

\item{path_folder}{The path where all results are stored.}

\item{splines_part}{Logical value to indicate whether there are splines
terms in the equation to draw the quantile contours.}

\item{wValue}{Fixed value to be plugged in the spline part of the equation.}

\item{print_plot}{Logical determining whether plot should be printed or
data with coordinantes should be returned. Only checked when paintedArea is
FALSE. Default is TRUE.}

\item{model_name}{When results will be collected in a folder, this should be
the name of the name considered by BayesX to save all tables. Default is
'bayesx.estim'.}

\item{name_var}{When there is a nonlinear variable from which one wants to
consider different values for plotting, this should have the name of the
variable.}

\item{...}{Other parameters for \code{summary.multBQR}.}
}
\value{
A ggplot with the quantile regions based on Bayesian quantile
 regression model estimates.
}
\description{
This function draws plots of the quantile region based on multiple-output
 quantile regression models, when the response variable has 3 dimensions.
}
