% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multBayesQR.R
\name{multBayesQR}
\alias{multBayesQR}
\title{Multiple-output Bayesian quantile regression model}
\usage{
multBayesQR(response, formulaPred, directionPoint, tau = 0.5, dataFile,
  itNum = 2000, burnin, thin = 1, betaValue = NULL, sigmaValue = 1,
  vSampleInit = NULL, priorVar = 100, hyperSigma = c(0.1, 0.1),
  refresh = 100, bayesx = TRUE, sigmaSampling = TRUE, quiet = T,
  tobit = FALSE, numCores = 1, recordLat = FALSE, outfile = NULL,
  check_bayesx = FALSE, path_bayesx = NULL, ...)
}
\arguments{
\item{response}{Names of response variables}

\item{formulaPred}{a formula object, with . on the left side of a ~ operator,
and the predictor terms, separated by + operators, on the right side.}

\item{directionPoint}{Either a vector with the same number of dimensions of
response variable, indicating a direction, or a integer indicating the
number of directions equally spaced in the unit circle one should
estimate. When the response has more than 2 dimensions, this should
determine the number of points considered in each dimension, in order to
define the number of directions taken into account in estimation. The
number of directions is then equal to \code{length_1d}^d, where d is the
dimension of the response variable.}

\item{tau}{Quantiles of interest. Default is th median, \code{tau = 0.5}.}

\item{dataFile}{A data.frame from which to find the variables defined in the
formula.}

\item{itNum}{Number of iterations.}

\item{burnin}{Size of the initial to be discarded.}

\item{thin}{Thinning parameter. Default value is 1.}

\item{betaValue}{Initial values for the parameter beta for the continuous
part.}

\item{sigmaValue}{Initial value for the scale parameter.}

\item{vSampleInit}{Initial value for the latent variables.}

\item{priorVar}{Value that multiplies a identity matrix in the elicition
process of the prior variance of the regression parameters.}

\item{hyperSigma}{Vector of size containing the hyperparameters of the
inverse gamma distribution for the sigma parameter of the asymmetric
Laplace distribution. Default is c(0.1, 0.1), which gives a noninformative
prior for sigma.}

\item{refresh}{Interval between printing a message during the iteration
process. Default is set to 100.}

\item{bayesx}{If TRUE, the default, it uses bayesX software to estimate
the quantile regression oarameters, which can be faster. If FALSE, it
uses a Rcpp implementation of the MCMC sampler.}

\item{sigmaSampling}{If TRUE, the default, it will sample from the posterior
distribution of the scale parameter. If FALSE, all values will be fixed to
1.}

\item{quiet}{If TRUE, the default, it does not print messages to check if
the MCMC is actually updating. If FALSE, it will use the value of refresh
to print messages to control the iteration process.}

\item{tobit}{If TRUE, it will input the censored value for all observations
with y = 0, according to the model. If FALSE, the default, it will estimate
the parameter without this inputation process.}

\item{numCores}{The number of cores that could be used for estimating
parallel models when more than one direction is considered.}

\item{recordLat}{If TRUE, it will keep the Markov chain samples for the
latent variable. Default is FALSE.}

\item{outfile}{argument to be passed to \code{bayesx.control}, in order
to define a directory where all output files should be saved.}

\item{check_bayesx}{To check whether all calls to BayesX generated valid
chain values for all models. In case there are NA values, it calls BayesX
just for those models with problems. This is only considered when
\code{outfile} is different than \code{NULL}.}

\item{path_bayesx}{When \code{check_bayes} is \code{TRUE}, the user must
inform the path of BayesX in order for these new calls of the program.}

\item{...}{arguments passed to \code{bayesx.control}.}
}
\value{
A list with the chains of all parameters of interest.
}
\description{
This function estimates a multiple-output Bayesian quantile regression model
}
