% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.summary.bqr.R, R/plot.summary.spBQR.R, R/plot.summary.zitobitQR.R
\name{plot.summary.bqr}
\alias{plot.summary.bqr}
\title{Plot methods for summary of Bayesian quantile regression models.}
\usage{
\method{plot}{summary.bqr}(object, separate = F, sigma = F, ...)

\method{plot}{summary.bqr}(object, separate = F, sigma = F, ...)

\method{plot}{summary.bqr}(object, separate = F, sigma = F, ...)
}
\arguments{
\item{object}{This is an object of the class "summary.bqr",
produced by a call to the summary.bqr function.}

\item{separate}{if FALSE, all plots are returned in the same output, and
if TRUE the plot for each variable is returned separately. Default value
is FALSE.}

\item{sigma}{if FALSE, it does the plot the posterior estimates for quantile
regression parameters, and if TRUE, it plots the posterior estimates only
for sigma. Default is set to FALSE.}

\item{object}{This is an object of the class "summary.bqrs",
produced by a call to the summary.bqr function.}

\item{separate}{if FALSE, all plots are returned in the same output, and
if TRUE the plot for each variable is returned separately. Default value
is FALSE.}

\item{sigma}{if FALSE, it does the plot the posterior estimates for quantile
regression parameters, and if TRUE, it plots the posterior estimates only
for sigma. Default is set to FALSE.}

\item{object}{This is an object of the class "summary.bqrs",
produced by a call to the summary.bqr function.}

\item{separate}{if FALSE, all plots are returned in the same output, and
if TRUE the plot for each variable is returned separately. Default value
is FALSE.}

\item{sigma}{if FALSE, it does the plot the posterior estimates for quantile
regression parameters, and if TRUE, it plots the posterior estimates only
for sigma. Default is set to FALSE.}
}
\value{
A ggplot of the posterior estimates with their credible intervals.

A ggplot of the posterior estimates with their credible intervals.

A ggplot of the posterior estimates with their credible intervals.
}
\description{
Returns a ggplot of the estimates of the models and their respective
 credible intervals.

Returns a ggplot of the estimates of the models and their respective
 credible intervals.

Returns a ggplot of the estimates of the models and their respective
 credible intervals.
}

