% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.summary.spBQR.R
\name{plot.summary.spBQR}
\alias{plot.summary.spBQR}
\title{Plot methods for summary of spatial Bayesian quantile regression models.}
\usage{
\method{plot}{summary.spBQR}(x, separate = FALSE, beta = TRUE,
  sigma = FALSE, lambda = FALSE, alpha = FALSE, ...)
}
\arguments{
\item{x}{This is an object of the class "summary.spBQR",
produced by a call to the summary.spBQR function.}

\item{separate}{if FALSE, all plots are returned in the same output, and
if TRUE the plot for each variable is returned separately. Default value
is FALSE.}

\item{beta}{If TRUE, it plots the posterior estimates for beta(tau).
Default is set to TRUE.}

\item{sigma}{If TRUE, it plots the posterior estimates for sigma.
Default is set to FALSE.}

\item{lambda}{If TRUE, it plots the posterior estimates for lambda.
Default is set to FALSE.}

\item{alpha}{If TRUE, it plots the posterior estimates for alpha.
Default is set to FALSE.}

\item{...}{other plot params (currently not used)}
}
\value{
A ggplot of the posterior estimates with their credible intervals.
}
\description{
Returns a ggplot of the estimates of the models and their respective
 credible intervals.
}
