% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCensoring.R
\name{plotCensoring}
\alias{plotCensoring}
\title{Plot function to compare the posterior distribution of the censoring
 indicator given a factor used in the model.}
\usage{
plotCensoring(object, variable)
}
\arguments{
\item{object}{This is an object of the class "zitobitQR", produced by a call
to the zitobitQR function.}

\item{variable}{Name of the variable to be used in the comparison.}
}
\value{
A plot of the posterior density of the probability of being censored
 given some factor used in the model.
}
\description{
Returns a plot comparing the posterior distribution for
}

