% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotComparison.R
\name{plotComparison}
\alias{plotComparison}
\title{Plot function to compare the posterior distribution for the latent variables}
\usage{
plotComparison(object, observation, burnin = 50, plotComp = T, ...)
}
\arguments{
\item{object}{This is an object of the class "bqr", produced by a call
to the bayesQR function.}

\item{observation}{Number of the observation to be compared with all the
other observations, in the posterior distribution of their latent
variables.}

\item{burnin}{Initial part of the chain, which is to be discarded. Default
value is 50.}

\item{plotComp}{If TRUE, the function prints the plot with the comparison.
Default is set to TRUE.}
}
\value{
A plot of the posterior density of the latent variable of a selected
 observation and the correspondent density for all other observations.
}
\description{
Returns a plot comparing the posterior distribution for an observation and
 the distribution for all other observations.
}

