% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotvfit.R
\name{plotvfit}
\alias{plotvfit}
\title{Plot function to analyse the posterior latent variables in a Bayesian
quantile regression model.}
\usage{
plotvfit(object, separate = F, ...)
}
\arguments{
\item{object}{This is an object of the class "bqr",
produced by a call to the bayesQR function.}

\item{separate}{if FALSE, all plots are returned in the same output, and
if TRUE the plot for each variable is returned separately. Default value
is FALSE.}
}
\value{
A ggplot of the posterior trace for all latent variables.
}
\description{
Returns a ggplot of the posterior trace of the latent variables
}

