% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.bqr.R
\name{summary.bqr}
\alias{summary.bqr}
\title{Summary methods for Bayesian quantile regression models.}
\usage{
\method{summary}{bqr}(object, burnin = 1000, ci = 0.95, mult = FALSE, ...)
}
\arguments{
\item{object}{This is an object of class "bqr", produced by a call to the
bayesQR function.}

\item{burnin}{Initial part of the chain, which is to be discarded. Default
value is 1000.}

\item{ci}{Credible interval coefficient. Default value is 0.95.}

\item{mult}{Defining whether this function will be used for multiple-output
Bayesian quantile regression or not. Default is FALSE.}

\item{...}{other summary parameters (not used)}
}
\value{
A data frame with summary information about the quantile regression
 parameters.
}
\description{
Returns a summary data.frame for a Bayesian quantile regression fit for more
 than one quantile.
}
\examples{
set.seed(1)
}
