% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summary.twopartQR.R
\name{summary.twopartQR}
\alias{summary.twopartQR}
\title{Summary methods for Bayesian quantile regression models.}
\usage{
\method{summary}{twopartQR}(object, burnin = 500, ci = 0.95, ...)
}
\arguments{
\item{object}{This is an object of class "twopartQR", produced by a call to the
twopartQR function.}

\item{burnin}{Initial part of the chain, which is to be discarded. Default
value is 1000.}

\item{ci}{Credible interval coefficient. Default is 0.95.}
}
\value{
A data frame with summary information about the quantile regression
 parameters.
}
\description{
Returns a summary data.frame for a Bayesian quantile regression fit for more
 than one quantile.
}
\examples{
set.seed(1)
}

