% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/update.twopartQR.R
\name{update.twopartQR}
\alias{update.twopartQR}
\title{Model update}
\usage{
\method{update}{twopartQR}(model, itNum, thin, sigmaGamma, ...)
}
\arguments{
\item{model}{Fitted model by the function twopartQR}

\item{thin}{Thinning parameter for the chains.}

\item{sigmaGamma}{Tuning parameter for the Metropolis-Hasting part of the
algorithm.}

\item{numIt}{Size of the chain that will be added to the current available
chain.}
}
\value{
A list with the updated chains.
}
\description{
This function updates the chains for a model fitted by the function
zitobitQR.
}
\examples{
set.seed(1)
}
\seealso{
\code{\link{twopartQR}}
}

