% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/update.zitobitQR.R
\name{update.zitobitQR}
\alias{update.zitobitQR}
\title{Model update}
\usage{
\method{update}{zitobitQR}(model, numIt, thin, sigmaGamma)
}
\arguments{
\item{model}{Fitted model by the function zitobitQR}

\item{numIt}{Size of the chain that will be added to the current available
chain.}

\item{thin}{Thinning parameter for the chains.}

\item{sigmaGamma}{Tuning parameter for the Metropolis-Hasting part of the
algorithm.}
}
\value{
A list with the updated chains.
}
\description{
This function updates the chains for a model fitted by the function
zitobitQR.
}
\examples{
set.seed(1)
}
\seealso{
zitobitQR
}

