% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesQR.R
\name{bayesQR}
\alias{bayesQR}
\title{Bayesian quantile regression model}
\usage{
bayesQR(formula, tau = 0.5, data, itNum, thin = 1, betaValue = NULL,
  sigmaValue = 1, vSampleInit = NULL, priorVar = 100, refresh = 100,
  sigmaSampling = TRUE, quiet = T, tobit = FALSE, recordLat = FALSE,
  blocksV = 0, stopOrdering = FALSE, numOrdered = itNum/2)
}
\arguments{
\item{formula}{a formula object, with the response on the left of a ~
operator, and the terms, separated by + operators, on the right.}

\item{tau}{Quantile of interest.}

\item{data}{A data.frame from which to find the variables defined in the
formula}

\item{itNum}{Number of iterations.}

\item{thin}{Thinning parameter. Default value is 1.}

\item{betaValue}{Initial values for the parameter beta for the continuous
part.}

\item{sigmaValue}{Initial value for the scale parameter.}

\item{vSampleInit}{Initial value for the latent variables.}

\item{priorVar}{Value that multiplies a identity matrix in the elicition
process of the prior variance of the regression parameters.}

\item{refresh}{Interval between printing a message during the iteration
process. Default is set to 100.}

\item{quiet}{If TRUE, the default, it does not print messages to check if
the MCMC is actually updating. If FALSE, it will use the value of refresh
to print messages to control the iteration process.}

\item{tobit}{If TRUE, it will input the censored value for all observations
with y = 0, according to the model. If FALSE, the default, it will estimate
the parameter without this inputation process.}

\item{recordLat}{If TRUE, it will keep the Markov chain samples for the
latent variable. Default is FALSE.}

\item{blocksV}{Number of blocks used to sample in the posterior distribution
of the latent variable. If 0, then blocking is not used and all latent
observations are sampled from. Default value is 0.}

\item{stopOrdering}{If TRUE, it will stop ordering the weighted residuals
in order to update the states of the latent variables, and will consider
the ordering of some particular state of the chain; if FALSE, for every
iteration of the MCMC procedure, it will keep reordering these residual
terms. Default is FALSE.}

\item{numOrdered}{The number of iterations that will be used to order
the weighted residuals needed for the update of the posterior
distribution of the latent variables. Default is half the size of
the MCMC chain.}
}
\value{
A list with the chains of all parameters of interest.
}
\description{
This function estimates a Bayesian quantile regression model
The response variable is assumed to follow a asymmetric Laplace
distribution.
}
\examples{
set.seed(1)
}
\references{
Kozumi and Kobayashi (2011) - Gibbs sampling methods for
 Bayesian quantile regression. Journal of Statistical Computation and
 Simulation.
}
