% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawQuantileRegion.R
\name{drawQuantileRegion}
\alias{drawQuantileRegion}
\title{Multiple-output Bayesian quantile regression model}
\usage{
drawQuantileRegion(model, ngridpoints = 100, xValue = 1,
  paintedArea = TRUE, comparison = FALSE, ...)
}
\arguments{
\item{model}{This is an object of the class \code{multBQR}, produced by a call to the
\code{multBayesQR} function.}

\item{ngridpoints}{Number of grid points considered to build this quantile region,
where a thorough search will look for the specified region, given the estimates
for several directions. Default is 100, that will produce a grid with 10.000 points
in the observed range of the data.}

\item{xValue}{Fixed value of the predictor variables. Default value is 1 when there is
only the intercept. If there is the interest in comparing the quantile regions
for different values of predictors, it must used with a list of values for the
predictors.}

\item{paintedArea}{If TRUE, it will plot the data points and the quantile region layer
over the points, for each tau, in different plots. If FALSE, it will produce one
plot showing showing all quantile regions for the different quantiles.}

\item{comparison}{Only considered when \code{paintedArea = FALSE}, if TRUE then
it will plot comparisons of quantile regions for different values of \code{xvalue}.}

\item{...}{Other parameters for \code{summary.multBQR}.}
}
\value{
A ggplot with the quantile regions based on Bayesian quantile regression
model estimates.
}
\description{
This function draws plots of the quantile region based on multiple-output quantile
regression models.
}
