% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawQuantileRegion.R
\name{drawQuantileRegion}
\alias{drawQuantileRegion}
\title{Multiple-output Bayesian quantile regression model}
\usage{
drawQuantileRegion(
  model,
  datafile,
  response,
  ngridpoints = 100,
  xValue = 1,
  paintedArea = FALSE,
  comparison = FALSE,
  result_folder = FALSE,
  path_folder = NULL,
  splines_part = FALSE,
  wValue = NULL,
  print_plot = TRUE,
  model_name = "bayesx.estim",
  name_var,
  range_y = NULL,
  lower_q = FALSE,
  upper_q = FALSE,
  lambda_a = NULL,
  ...
)
}
\arguments{
\item{model}{This is an object of the class \code{multBQR}, produced by a
call to the \code{multBayesQR} function.}

\item{datafile}{A data.frame from which to find the variables defined in the
formula.}

\item{response}{Names of response variables}

\item{ngridpoints}{Number of grid points considered to build this quantile
region, where a thorough search will look for the specified region, given
the estimates for several directions. Default is 100, that will produce a
grid with 10.000 points in the observed range of the data.}

\item{xValue}{Fixed value of the predictor variables. Default value is 1
when there is only the intercept. If there is the interest in comparing the
quantile regions for different values of predictors, it must used with a
list of values for the predictors.}

\item{paintedArea}{If TRUE, it will plot the data points and the quantile
region layer over the points, for each tau, in different plots. If FALSE,
it will produce one plot showing showing all quantile regions for the
different quantiles.}

\item{comparison}{Only considered when \code{paintedArea = FALSE}, if TRUE
then it will plot comparisons of quantile regions for different values of
\code{xvalue}.}

\item{result_folder}{Logical value determining whether all estimates are
stored in a folder produced by a BayesX call. Default is FALSE.}

\item{path_folder}{If \code{result_folder = TRUE}, then one must inform the
path where all results are stored.}

\item{splines_part}{Logical value to indicate whether there are splines
terms in the equation to draw the quantile contours.}

\item{wValue}{Fixed value to be plugged in the spline part of the equation.}

\item{print_plot}{Logical determining whether plot should be printed or
data with coordinantes should be returned. Only checked when paintedArea is
FALSE. Default is TRUE.}

\item{model_name}{When results will be collected in a folder, this should be
the name of the name considered by BayesX to save all tables. Default is
'bayesx.estim'.}

\item{name_var}{When there is a nonlinear variable from which one wants to
consider different values for plotting, this should have the name of the
variable.}

\item{range_y}{matrix type object containing in the first line the range for
the first dimension and in the second line the range for the second
dimension of Y. This will be used to find the respective quantile regions.}

\item{lower_q}{Logical determining whether quantile region should be based
on the lower value of 0.95 credible interval for each parameter of the model.}

\item{upper_q}{Logical determining whether quantile region should be based
on the upper value of 0.95 credible interval for each parameter of the model.}

\item{lambda_a}{Parameter to be make the adjustment of the intercept to
correct the miscoverage of the quantile regions.
\code{lower_q} must be FALSE for this argument to be checked.}

\item{...}{Other parameters for \code{summary.multBQR}.}
}
\value{
A ggplot with the quantile regions based on Bayesian quantile
 regression model estimates.
}
\description{
This function draws plots of the quantile region based on multiple-output
 quantile regression models.
}
