% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/latentKL.R
\name{latentKL}
\alias{latentKL}
\title{Probability of being an outlier given the Bayesian quantile regression fits.}
\usage{
latentKL(object, burnin = 50, plotKL = TRUE, scales.free = FALSE,
  all.obs = TRUE, obs)
}
\arguments{
\item{object}{This is an object of the class "bqr", produced by a call
to the bayesQR function.}

\item{burnin}{Initial part of the chain, which is to be discarded. Default
value is 50.}

\item{scales.free}{If FALSE, the default, then all plots will use the same
y scale. If TRUE, for each tau the plot will use the best possible scale
in order to visualize the probability information for all observations.}

\item{plotProb}{If TRUE, the function prints the plot with all probabilities.
Default is set to TRUE.}
}
\value{
Prints a plot of the posterior probability of being an outlier for
 all observations and returns a data.frame with all values of the
 probabilities.
}
\description{
Returns the probability that each observation is an outlier given the
 Bayesian quantile regression fits.
}

