% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probOutlier.R
\name{probOutlier}
\alias{probOutlier}
\title{Probability of being an outlier given the Bayesian quantile regression fits.}
\usage{
probOutlier(object, burnin = 50, plotProb = TRUE, scales.free = FALSE,
  all.obs = TRUE, obs)
}
\arguments{
\item{object}{This is an object of the class "bqr", produced by a call
to the bayesQR function.}

\item{burnin}{Initial part of the chain, which is to be discarded.
Default value is 50.}

\item{plotProb}{If TRUE, the function prints the plot with all probabilities.
Default is set to TRUE.}

\item{scales.free}{If FALSE, the default, then all plots will use the same
y scale. If TRUE, for each tau the plot will use the best possible scale
in order to visualize the probability information for all observations.}

\item{all.obs}{if TRUE, calculates KL divergence for all observations}

\item{obs}{if \code{all.obs} is FALSE, specifies the observation to
calculate the KL divergence.}
}
\value{
Prints a plot of the posterior probability of being an outlier for
 all observations and returns a data.frame with all values of the
 probabilities.
}
\description{
Returns the probability that each observation is an outlier given the
 Bayesian quantile regression fits.
}
