% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summary.bqr.R, R/summary.spBQR.R
\name{summary.bqr}
\alias{summary.bqr}
\title{Summary methods for Bayesian quantile regression models.}
\usage{
\method{summary}{bqr}(object, burnin = 1000, ci = 0.95, ...)

\method{summary}{bqr}(object, burnin = 1000, ci = 0.95, ...)
}
\arguments{
\item{object}{This is an object of class "bqrs", produced by a call to the
bayesQR function.}

\item{burnin}{Initial part of the chain, which is to be discarded. Default
value is 1000.}

\item{object}{This is an object of class "bqrs", produced by a call to the
bayesQR function.}

\item{burnin}{Initial part of the chain, which is to be discarded. Default
value is 1000.}
}
\value{
A data frame with summary information about the quantile regression
 parameters.

A data frame with summary information about the quantile regression
 parameters.
}
\description{
Returns a summary data.frame for a Bayesian quantile regression fit for more
 than one quantile.

Returns a summary data.frame for a Bayesian quantile regression fit for more
 than one quantile.
}
\examples{
set.seed(1)
set.seed(1)
}

