% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/update.bqr.R
\name{update.bqr}
\alias{update.bqr}
\title{Model update for a Bayesian quantile regression model}
\usage{
\method{update}{bqr}(model, itNum, ...)
}
\arguments{
\item{model}{Fitted model by the function bayesQR}

\item{numIt}{Size of the chain that will be added to the current available
chain.}

\item{thin}{Thinning parameter for the chains.}
}
\value{
A list with the updated chains.
}
\description{
This function updates the chains for a model fitted by the function bayesQR.
}
\examples{
set.seed(1)
}
\seealso{
\code{\link{twopartQR}}
}

