% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zitobitQR.R
\name{zitobitQR}
\alias{zitobitQR}
\title{Bayesian quantile regression model with a discrete component at zero}
\usage{
zitobitQR(tau = 0.5, y, X, itNum, thin = 1, betaValue = rep(0, dim(X)[2]),
  sigmaValue = 1, gammaValue = rep(0, dim(X)[2]), sigmaGamma = 0.5,
  link = 1)
}
\arguments{
\item{tau}{Quantile of interest.}

\item{y}{Vector of response variable.}

\item{X}{Design matrix.}

\item{itNum}{Number of iterations.}

\item{thin}{Thinning parameter.}

\item{betaValue}{Initial values for the parameter beta for the continuous part.}

\item{sigmaValue}{Initial value for the scale parameter.}

\item{link}{Integer defining the link function used for the probability
model. Default is 1.
for the logit link function.}

\item{gamaValue}{Initial value for the parameter gamma of the discrete part.}

\item{sigmaBetaZero}{Tuning parameter for the Metropolis-Hastings step.}
}
\value{
A list with the chains of all parameters of interest.
}
\description{
This function estimates a bayesian quantile regression model with a discrete component
at zero, where all zero observations are assumed to distributed according to a mixed
discrete-continuous distribution.
}
\examples{
set.seed(1)
}
\references{
Santos and Bolfarine (2015) - Bayesian quantile regression analysis for continuous data
with a discrete component at zero. \emph{Preprint}. \url{http://arxiv.org/abs/1511.05925}
}

